config/*: refactor config.go's Parse() to use GetConfigVersion
